Add a timeout in ym2610_write_port0
The non-psub version of ym2610_write_port0 could infinitely loop if we never
see the ym2610 as having unset its busy bit on YM2610_PORT0_REGISTER. The
only function that calls this is ym2610_timer_init which is called when
testing both getting irqs from the ym2610 and not getting them. A timeout
was added to ym2610_write_port0 to avoid the infinite loop. As such 2 new
errors were created:

0x18 = YM2610 TIMER INIT (NOIRQ TEST)
0x19 = YM2610 TIMER INIT (IRQ TEST)

To identify the timeout condition occured and which one.

This resolves a stall if the z80's ram OE was stuck low. For whatever reason
on my MV1T test hardware this was not causing a data bus conflict between ram
and rom, but when trying to setup the ym2610 timer it would trigger the
infiniate loop issue.

In testing this I also found there is a risk of test functions after the
68k/z80 comm test overwritng the z80 -> 68k ACK message before the 68k had
time to read it. A small delay was added after sending the ACK and also
tightened the loop to read the ACK.