Add sm1 crc32 and output enable tests
These tests are accomplisted by having the diag m1 run code out of ram while the
diag bios handles swapping to/from the sm1.  The sm1 tests will only be allowed
to run if a slot switch was done to activate the diag m1 and the 68k <=> Z80
test passed.

The sm1 crc32 test only checks the first 0x5400 bytes of the rom, which covers
all of the actual code in the sm1 rom.  The expected crc32 for that  data range
is 0xbd94a5a6.

If sm1 tests are run "[SM1]" text will be printed in the top right of the
screen to the left of "[SSx]" slot switch indicator.