Rework how details about error codes get printed
Previously the middle 4 bits of the error code (byte) were used to determine which print function should be called for the error code.  Because of this unassigned error codes already have specific error print functions they will use.  This was making it impossible to add all the necessary error codes for the coming memory card tests, as there aren't enough free error codes with the needed print functions.

There was a pad/alignment byte in each error code struct of the EC_LOOKUP_TABLE table.  This extra byte is now being used to identify the print function that should be called for the given error code.  Thus allowing you to set the print function for each error code.  Also did a little cleanup by moving all the print error related code to the same section in the source.

There should be no end user visual changes from this update.