Increase available bios code size from 16K to 32K
Running out of space for code so increasing the available code size to 32K.  Making this increase means A13 is now allocated to bios code and is no longer valid for the bios mirror test.  Likewise the location of the bios crc and mirror byte have been adjusted for the new size.  Docs have been updated to reflect these changes.