replace all xyp instances with xp, nuke xyp and print16 code.
This replaces all xyp strings with xy strings and remove all xyp code.  The p in xyp was for setting the palette when writing the string to the fix layer.  However every instance was just setting it to 0, which is the same thing that happens when printing an xy string.

Also remove some dead print16 functions.