dynamic subroutines
Introduce dynamic subroutines (dsub) which are an expansion of the original 0.19 diag bios's pseudo subroutines (psub) concept.  dsubs have 2 modes of operation.  The original psub mode and a new real mode (rsub) mode, which allows using the dsub code blocks like a real subroutine call (bsr/rts).  In rsub mode the return address is pushed onto the stack, instead of being stored in a4/a5/a7 like with psub.   Likewise when a dsub needs to return, in rsub mode an rts is done instead of jumping to a4/a5/a7 like psub's do.  More specific details can be found in the large comment block near the top of sp1.asm.

The purpose of this is to reduce code duplication as a number of utility functions have 2 copies, one for psub and one for bsr/rts.