Add Simple|Small Subroutine A3 (SSA3), implement fix_seek_xy_ssa3
Simple|Small Subroutine A3 (SSA3)

One of the limitations with dsubs is only being able to nest 2 times.  This is mainly a problem with the print related dsubs as most will want to either seek to an x,y location in the fix layer and/or clearing a fix layer line. If print dsub were to dsub call fix_seek_xy|fix_clear_line it would often be one to many nests, so a bunch of the print dsub's had a copy of the code for fix_seek_xy|fix_clear_line in them.  SSA3 is meant to get around this by calling a simple|small subroutine that will always jmp (a3) to return.  This way we can get around having another dsub nest.

Created fix_seek_xy_ssa3 subroutine and updated a bunch of the print functions to use it.