Fix bug related to sp1 auto detection of diag m1
If the user held down D button to force diag m1 tests it would follow a code
path that would cause a cmp.b of REG_SOUND against an uninitialized d1.
If these ended up be equal, the bios would skip doing the slot switch.  This
would leave the board/bios in a state where it was trying to do the comm test
against the sm1.