Use saved P1 input after failed prepare for slot switch
Previously when the prepare for slot switch request failed we would
prompt the user to press start to force the slot switch.  If the
user was wanting to switch to slot 2+ they would also have to be
holding the correct direction on the joystick when they pressed start.
This unintuitive, so instead save the users input prior to the prepare
for slot switch request and use that when picking which slot to switch
to.