Implement error addresses
If an error occurs during automatic tests and you have the 'A'
button held down it will cause the diag rom to jump to a specific
address based on the error code and go into a tight loop. The
address it jumps to is

0xc06000 | (error_code << 5)

From this you can use a logic probe on the 68k's address lines
to figure out the error code.  This is specifically useful when
your video output is corrupt/dead and you don't have a working
diag m1 cart/sound to provide beep codes.

There is additional information about this in docs/error_codes.md