sp1: menu cleanup/rework
Instead of having a hardcoded main menu, convert it into a generic
menu handler that you pass in a list of menu entries.  It will
handle displaying, navigation and running the select menu entry's
code.  This should also allow for having sub menu's off the main
menu.

At somepoint in the future I will add a memory viewer on the main
menu that will goto a sub menu that lets you pick which memory
region to start the memory viewer at.