Change how versioning works
Instead of using 0.19aXX versions, switch to using the last commit date. Such
that the header will now be

NEO DIAG - SMKDAN/ACK - YYYY-MM-DD

If the diag is compiled when there are unchecked in changes the date will be
appended with the letter "D" to indicate its a dirty build (ie: 2025-02-27D)

Add 'DIAG GIT HASH' to the debug menu, which will show the shortened git hash.
If the build is dirty, the word "DIRTY" will appear after the hash value.

Lastly this version string is also compiled into the diag m1 rom, making it
possible to see what version it is by looking at it in a hex editor or
running `strings` on it.

```
$ strings m1/build/m1.bin
{/Oz/G}/_|/W
))))x
))))x
B       *
BDM!
R       !
8X>'
))))>0
))))
5A8CD75 DIRTY
NEO DIAG - SMKDAN/ACK - 2025-02-27D
...
```